Extract ConfirmModal as standalone component
