+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/InteractionLayer/components/SubTaskPopup/components/ConfirmModal/locales/en.json
@@ -1,0 +1,7 @@
+{
+  "Confirm": {
+    "close": "Close without saving",
+    "confirm": "Are you sure you want to close? This mark will be deleted.",
+    "keepWorking": "Keep working"
+  }
+}
