+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/InteractionLayer/components/SubTaskPopup/locales/en.json
@@ -1,8 +1,5 @@
 {
   "Task": {
-    "close": "Close without saving",
-    "confirm": "Are you sure you want to close? This mark will be deleted.",
-    "keepWorking": "Keep working",
     "notRender": "Task component could not be rendered.",
     "required": "This task is required."
   }
